requirements.txt
pytest==8.3.3
httpx==0.27.2
python-dotenv==1.0.1
allure-pytest==2.13.5
